+++ src/main/java/com/codetaylor/mc/athenaeum/network/IPacketService.java
@@ -9,6 +9,10 @@ import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
  */
 public interface IPacketService {
 
+  void sendToAllAround(IMessage message, int dimension, double x, double y, double z, double range);
+
+  void sendToAllAround(IMessage message, int dimension, double x, double y, double z);
+
   void sendToAllAround(IMessage message, TileEntity tileEntity, int range);
 
   void sendToAllAround(IMessage message, TileEntity tileEntity);